% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{pumpkins_palette}
\alias{pumpkins_palette}
\title{Colors from the Smashing Pumpkins Album Covers for data visualization}
\usage{
pumpkins_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of album cover to extract color palette from. Thus far, the following albums are available:
\code{gish}, \code{gish_dark},  \code{gish_light},
\code{simamese_dream}, \code{simamese_dream_dark},  \code{simamese_dream_light}, \code{simamese_dream_re},
\code{mellon_collie},  \code{mellon_collie_dark} , \code{mellon_collie_light} ,
\code{adore}, \code{zeitgeist}, \code{oceania}, \code{cyr}}

\item{n}{Number of colors desired.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colors.
@importFrom graphics rgb rect par image text}
}
\value{
A vector of colors.
}
\description{
Modeled after karthik/wesanderson (https://github.com/karthik/wesanderson), this
package contains color schemes used in the most iconic album covers by The Smashing Pumpkins.
}
\examples{

pumpkins_palette("gish")
pumpkins_palette("mellon_collie")
pumpkins_palette("oceania")
pumpkins_palette("simamese_dream_re", 4)

# If more colors than are saved in a palette are needed, please specify a
# continuous palette to interpolate between saved colors.

pal <- pumpkins_palette(21, name = "oceania", type = "continuous")
image(volcano, col = pal)

}
\keyword{colors}
