% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{pumpkins_pallettes}
\alias{pumpkins_pallettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 14.
}
\usage{
pumpkins_pallettes
}
\description{
Use \code{\link{pumpkins_palette}} to extract color palettes for a given length.
}
\keyword{datasets}
