% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmashingPumpkins.R
\docType{package}
\name{SmashingPumpkins}
\alias{SmashingPumpkins}
\title{SmashingPumpkins}
\description{
list of color palettes for data visualization
}
\details{
color palettes extracted from iconic 'The Smashing Pumpkins' album covers
}
